-changed mark count property type from int to double.

+++ src/Application/ZMK.Application/Contracts/MarkAddDTO.cs
@@ -1,3 +1,3 @@
 ﻿namespace ZMK.Application.Contracts;
 
-public record MarkAddDTO(Guid ProjectId, string Code, string Title, int Order, double Weight, int Count, string? Remark);
+public record MarkAddDTO(Guid ProjectId, string Code, string Title, int Order, double Weight, double Count, string? Remark);
