+++ src/Presentation/ZMK.Wpf/ViewModels/MarkViewModel.cs
@@ -28,7 +28,7 @@ public partial class MarkViewModel : ModifiableViewModel<MarkViewModel>
     [ObservableProperty]
     [NotifyPropertyChangedFor(nameof(UpdatableSign))]
     [NotifyPropertyChangedFor(nameof(TotalWeight))]
-    private int _count;
+    private double _count;
 
     public double TotalWeight => Weight * Count;
